Update state and font family
