+++ src/components/AppleContainer.js
@@ -1,14 +1,22 @@
 import React, { useState, useEffect } from 'react';
-import { useSelector } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
+import { fetchAppleData } from '../redux/apple/appleSlice';
 import DataHandler from './DataHandler';
 import Header from './Header';
 
 const Apple = () => {
   const { apple, isLoading, error } = useSelector((state) => state.apple);
+  const dispatch = useDispatch();
   const [isSearchVisible, setIsSearchVisible] = useState(false);
   const [searchQuery, setSearchQuery] = useState('');
   const [filteredData, setFilteredData] = useState([]);
   const [resultCount, setResultCount] = useState(0);
+
+  useEffect(() => {
+    if (apple.length === 0) {
+      dispatch(fetchAppleData());
+    }
+  }, [dispatch, apple]);
 
   useEffect(() => {
     if (searchQuery !== '') {
@@ -43,14 +51,14 @@ const Apple = () => {
     <>
       <Header handleSearch={handleSearch} />
       {isSearchVisible && (
-        <div>
-          <input
-            type="text"
-            placeholder="Search by year..."
-            value={searchQuery}
-            onChange={handleSearchChange}
-          />
-        </div>
+      <div>
+        <input
+          type="text"
+          placeholder="Search by year..."
+          value={searchQuery}
+          onChange={handleSearchChange}
+        />
+      </div>
       )}
       <p className="prev-available">
         Previous Years Available:
