+++ src/components/GoogleContainer.js
@@ -1,14 +1,22 @@
 import React, { useEffect, useState } from 'react';
-import { useSelector } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
+import { fetchGoogleData } from '../redux/google/googleSlice';
 import DataHandler from './DataHandler';
 import Header from './Header';
 
 const Google = () => {
   const { google, isLoading, error } = useSelector((state) => state.google);
+  const dispatch = useDispatch();
   const [isSearchVisible, setIsSearchVisible] = useState(false);
   const [searchQuery, setSearchQuery] = useState('');
   const [filteredData, setFilteredData] = useState([]);
   const [resultCount, setResultCount] = useState(0);
+
+  useEffect(() => {
+    if (google.length === 0) {
+      dispatch(fetchGoogleData());
+    }
+  }, [dispatch, google]);
 
   useEffect(() => {
     if (searchQuery !== '') {
@@ -43,14 +51,14 @@ const Google = () => {
     <>
       <Header handleSearch={handleSearch} />
       {isSearchVisible && (
-        <div>
-          <input
-            type="text"
-            placeholder="Search by year..."
-            value={searchQuery}
-            onChange={handleSearchChange}
-          />
-        </div>
+      <div>
+        <input
+          type="text"
+          placeholder="Search by year..."
+          value={searchQuery}
+          onChange={handleSearchChange}
+        />
+      </div>
       )}
       <p className="prev-available">
         Previous Years Available:
